feat: use coordinax.operators
