+++ src/galax/coordinates/operators/__init__.py
@@ -3,20 +3,10 @@
 E.g. a translation.
 """
 
-from . import base, composite, funcs, galilean, identity, rotating, sequential
-from .base import *
-from .composite import *
-from .funcs import *
-from .galilean import *
-from .identity import *
-from .rotating import *
-from .sequential import *
+from . import _compat, _rotating
+from ._compat import *
+from ._rotating import *
 
 __all__: list[str] = []
-__all__ += base.__all__
-__all__ += composite.__all__
-__all__ += sequential.__all__
-__all__ += identity.__all__
-__all__ += galilean.__all__
-__all__ += rotating.__all__
-__all__ += funcs.__all__
+__all__ += _compat.__all__
+__all__ += _rotating.__all__
